correccion de error de componente y separacion de logica de rest-password
